Serialization and deserialization
two classes: Car implements Serializable and Engine does not.
Car has a reference field of type Engine.
We implement writeObject() to serialize an Engine object by manually serializing all of its fields separately. We also serialize the serializable String field.
We implement readObject() to deserialize a Car object by reading the double and int fields that are a part of Engine and creating a new Engine object with the provided constructor.